requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
django-modeltranslation==0.18.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
